Split API/growatt.py into client/device/detection
growatt.py mixed three concerns in one 470-line file: the Modbus transport,
the device abstraction, and device detection (whose module-level
get_device_info even shadowed the GrowattModbusBase.get_device_info method).

Split into focused modules:
- API/client.py   - GrowattModbusBase + GrowattSerial/GrowattNetwork transport
- API/device.py   - GrowattDevice + get_register_information
- API/detection.py - the device-probe get_device_info

Also replace the get_register_information if/elif chain with a _REGISTER_SETS
table lookup that lists each device type's holding registers, its (ordered)
input-register sets, and max read length. Behaviour is unchanged.

Update the two importers (__init__, config_flow) and add
tests/test_register_sets.py asserting the layout chosen for every device type
(incl. the hybrid input merge). Adds pymodbus to requirements_test.txt since
the device module imports the transport.
