+++ custom_components/growatt_modbus/__init__.py
@@ -32,7 +32,8 @@ from homeassistant.util import dt as dt_util
 from .API.device_type.base import GrowattDeviceRegisters
 from .API.utils import RegisterKeys
 from .API.const import DeviceTypes
-from .API.growatt import GrowattDevice, GrowattSerial, GrowattNetwork
+from .API.client import GrowattSerial, GrowattNetwork
+from .API.device import GrowattDevice
 from .const import (
     CONF_LAYER,
     CONF_SERIAL,
