+++ requirements_test.txt
@@ -1,1 +1,2 @@
 pytest>=8.0
+pymodbus>=3.8.3
